Add todomvc demo to the showcase
