+++ examples/showcase/Cargo.toml
@@ -16,3 +16,4 @@ mount_point = { path = "sub/mount_point" }
 npm_and_rest = { path = "sub/npm_and_rest" }
 textarea = { path = "sub/textarea" }
 timer = { path = "sub/timer" }
+todomvc = { path = "sub/todomvc" }
